requirements.txt
numpy
flask~=2.1.1
views~=0.3
setuptools~=65.6.3
six~=1.16.0
requests~=2.28.2
soundfile~=0.12.1
werkzeug~=2.2.3
connexion~=2.14.2
pyyaml~=6.0
pika~=1.3.1
openai~=0.27.2
torch~=2.0.0
diffusers~=0.14.0
pillow==9.4.0
pymongo~=4.3.3
transformers~=4.27.3
python-dotenv~=1.0.0
tqdm~=4.65.0